requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
python-dotenv